\name{gcipdr-package}
\alias{gcipdr-package}
\alias{gcipdr}
\docType{package}
\title{
\packageTitle{gcipdr}
}
\description{
\packageDescription{gcipdr}
}

\details{
This package is a beta version and extensive editing, documenting, and
amelioration shall be expected in the future.
}

\references{
 
}

\keyword{ package }
\keyword{ copula,  }
\seealso{
 \code{\link[<pkg>:<pkg>-package]{https://github.com/bonorico/gcipdr}} 
}
\examples{

## donn't run
## DataRebuild()


}
