% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipd_rec.R
\name{Return.IPD.design.matrix}
\alias{Return.IPD.design.matrix}
\title{Conversion of original IPD into suitable format for further processing.}
\usage{
Return.IPD.design.matrix(data, fill.missing = F)
}
\arguments{
\item{data}{an IPD.}

\item{fill.missing}{logical: if TRUE, application of `Return.key.IPD.summaries()` on the output of this function will instruct to compute IPD summaries such that `DataRebuild()` also generates imputations of the missing IPD records.}
}
\description{
At the level of a raw data holder this function converts the original IPD into a design-matrix-like format onto which marginal moments and pairwise correlations, needed by `DataRebuild()`, are computed.
}
\details{
See Details of `DataRebuild()` for more info on the required IPD format.
}
\seealso{
[Return.key.IPD.summaries()] for IPD summary computation, [DataRebuild()]
}
