% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipd_rec.R
\name{Return.key.IPD.summaries}
\alias{Return.key.IPD.summaries}
\title{Computation of input IPD summaries from the original IPD}
\usage{
Return.key.IPD.summaries(md, corrtype = c("rank.corr", "moment.corr",
  "normal.corr"), only.fixed.stats = FALSE, na.erase = TRUE)
}
\description{
This function computes marginal moments up to fourth degree, all pairwise correlation of an available IPD and Johnson system parameters for all IPD non-binary marginals. The IPD is supposed to be in a specific format (see `Return.IPD.design.matrix()` and Details of `DataRebuild()`).
}
\details{
Ideally the the IPD holder uses this function output as a template for the IPD summaries format to be disclosed and needed in `DataRebuild()` by a IPD summary holder. If IPD has missing data, the module can compute IPD summaries as if missing records were imputed. This must be decided before via argument fill.missing in `Return.IPD.design.matrix()`.
}
\seealso{
[Return.IPD.design.matrix()] for currently requested IPD format, [DataRebuild()]
}
