% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cppWrapper.R
\name{rsearch2}
\alias{rsearch2}
\title{Imposition of incomplete IPD correlation structure.}
\usage{
rsearch2(xell, x, refstat, mc.extremal.corr = NULL, dz = 0.001,
  random = TRUE, rf.type = c("joint.rank", "joint.moment", "rank.corr",
  "moment.corr", "tau.corr"), ties = c("min", "average", "max"))
}
\description{
`rsearch2()` implements a permutation-based re-ordering of first-degree-correlated IPD marginals only w.r.t. some fixed reference variable.
}
\details{
This function also wraps the C++ module 'findzerocorr' (see source code). This function is *not* a NORTA approach and is primarily meant for experimental and comparative purposes w.r.t. the 'complete-correlation approach' implemented in the NORTA transformation.
}
