% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipd_rec.R
\name{Simulate.many.datasets}
\alias{Simulate.many.datasets}
\title{Reconstruction of several (unrelated) available IPDs.}
\usage{
Simulate.many.datasets(datalist, H, method, fill.missing = FALSE,
  SBjohn.correction = F, stochastic.integration = F, checkdata = F,
  compute.eec = F, tabulate.similar.data = FALSE,
  set.corr.matr2null = FALSE)
}
\description{
This function can be used as simple application of `Simulate.data.given.IPD()` on a list of IPDs, DATA_1, DATA_2, DATA_3...
}
\seealso{
[Simulate.data.given.IPD()], [DataRebuild()]
}
