% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipd_rec.R
\name{Simulate.data.given.IPD}
\alias{Simulate.data.given.IPD}
\title{IPD reconstruction from IPD summaries directly computed on IPD.}
\usage{
Simulate.data.given.IPD(data, H = NULL, method, fill.missing = F,
  SBjohn.correction = F, stochastic.integration = F, checkdata = F,
  compute.eec = F, tabulate.similar.data = FALSE,
  print.message = TRUE, set.corr.matr2null = FALSE)
}
\description{
`Simulate.data.given.IPD()` is basically a wrapper for `DataRebuild()`
but additionally computes the needed input IPD summaries from the IPD automatically.
}

\section{Note}{
 the arguments of this function are mostly inherited by
 `DataRebuild()`. However fill.missing (logical, default to FALSE) is
 inherited by `Return.key.IPD.summaries()`. Also argument method
 (integer 1 to 4) is an encoding for different combinations of data.rearrange, corrtype, and marg.model (see 'setting.comb.matrix()' in source code).
}

\seealso{
[Return.key.IPD.summaries()] for currently allowed input IPD summary format, [DataRebuild()]
}
